back porting , Fix return type of ConceptService.getCountOfConceptReferenceTerms - TRUNK-2993
git-svn-id: http://svn.openmrs.org/openmrs/branches/1.9.x@25055 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/db/ConceptDAO.java
@@ -601,7 +601,7 @@ public interface ConceptDAO {
 	/**
 	 * @see ConceptService#getCountOfConceptReferenceTerms(String, ConceptSource, boolean)
 	 */
-	public Integer getCountOfConceptReferenceTerms(String query, ConceptSource conceptSource, boolean includeRetired)
+	public Long getCountOfConceptReferenceTerms(String query, ConceptSource conceptSource, boolean includeRetired)
 	        throws DAOException;
 	
 	/**
